move integration tests into single binary for optimized runtime
